requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
apache-airflow==2.7.3
apache-airflow-providers-docker==3.8.1
pyspark==3.5.0
requests==2.31.0
unidecode==1.3.7
pydantic>=1.10.0,<2.0.0
